+++ StoreRotationConfig/Patches/SyncShipUnlockablesPatches.cs
@@ -1,7 +1,10 @@
 using GameNetcodeStuff;
 using HarmonyLib;
 using StoreRotationConfig.Compatibility;
-using Unity.Netcode;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection.Emit;
+using UnityEngine;
 
 namespace StoreRotationConfig.Patches
 {
@@ -16,27 +19,121 @@ namespace StoreRotationConfig.Patches
         /// </summary>
         public static bool UnlockablesSynced { get; private set; } = false;
 
-        [HarmonyPatch(typeof(StartOfRound), nameof(StartOfRound.SyncShipUnlockablesClientRpc))]
-        [HarmonyPriority(Priority.Last)]
-        [HarmonyPostfix]
-        private static void SyncShipUnlockablesClientPost()
+        /// <summary>
+        ///     Fills 'storedItems' parameter immediately before 'SyncShipUnlockablesClientRpc()' is called on the server.  
+        /// </summary>
+        ///     ... (StartOfRound:1663)
+        ///     if(this.attachedVehicle != null)
+        ///     {
+        ///         vehicleID = this.attachedVehicle.vehicleID;
+        ///     }
+        ///     this.SyncShipUnlockablesClientRpc(playerSuitIDs, this.shipRoomLights.areLightsOn, placeableObjectPositions.ToArray(),
+        ///         placeableObjectRotations.ToArray(), placeableObjects.ToArray(),
+        ///         // storedItems.ToArray(),
+        ///         -> Delegate(),
+        ///         scrapValues.ToArray(), itemSaveData.ToArray(), vehicleID);
+        ///     )
+        /// <param name="instructions">Iterator with original IL instructions.</param>
+        /// <returns>Iterator with modified IL instructions.</returns>
+        [HarmonyPatch(typeof(StartOfRound), nameof(StartOfRound.SyncShipUnlockablesServerRpc))]
+        [HarmonyTranspiler]
+        private static IEnumerable<CodeInstruction> SyncShipUnlockablesServerRpcTranspiler(IEnumerable<CodeInstruction> instructions)
         {
-            // Return if local game instance is hosting the server.
-            if (NetworkManager.Singleton.IsHost || NetworkManager.Singleton.IsServer)
+            return new CodeMatcher(instructions).End().MatchBack(true,
+                new(OpCodes.Ldarg_0),
+                new(OpCodes.Ldloc_0),
+                new(OpCodes.Ldarg_0))
+            .Advance(10)
+            .SetInstructionAndAdvance(Transpilers.EmitDelegate((List<int> storedItems) =>
+                {
+                    // Probably not needed; list is always empty (hence the need for this).
+                    storedItems.Clear();
+
+                    // Actually add the IDs of every item currently in storage to the 'storedItems' list.
+                    for (int i = 0; i < StartOfRound.Instance.unlockablesList.unlockables.Count; i++)
+                    {
+                        if (StartOfRound.Instance.unlockablesList.unlockables[i].inStorage)
+                        {
+                            storedItems.Add(i);
+                        }
+                    }
+
+                    // Return as array for the 'SyncShipUnlockablesClientRpc()' call.
+                    return storedItems.ToArray();
+                }
+            )).InstructionEnumeration();
+        }
+
+        [HarmonyPatch(typeof(StartOfRound), nameof(StartOfRound.SyncShipUnlockablesServerRpc))]
+        [HarmonyPostfix]
+        private static void SyncShipUnlockablesServerRpcPost(StartOfRound __instance)
+        {
+            // Return if local client is not hosting the server.
+            if (!__instance.IsHost)
             {
                 return;
             }
 
-            // Return if unlockables are already synced with the host; toggle unlockable sync status if not.
-            if (UnlockablesSynced)
-            {
-                Plugin.StaticLogger?.LogDebug("Purchased unlockable items already synced with the host.");
+            // Manually trigger a store rotation on the server.
+            Plugin.Terminal?.RotateShipDecorSelection();
+        }
 
+        [HarmonyPatch(typeof(StartOfRound), nameof(StartOfRound.SyncShipUnlockablesClientRpc))]
+        [HarmonyPostfix]
+        private static void SyncShipUnlockablesClientRpcPost(StartOfRound __instance, UnlockablesList ___unlockablesList, int[] placeableObjects, int[] storedItems)
+        {
+            // Return if local client is hosting the server.
+            if (__instance.IsHost)
+            {
                 return;
             }
-            UnlockablesSynced = true;
 
-            // Manually trigger a store rotation.
+            // Check if unlockables need to be synced.
+            if (!UnlockablesSynced)
+            {
+                // Sync purchased suits.
+                Resources.FindObjectsOfTypeAll<UnlockableSuit>().Select(suit => ___unlockablesList.unlockables[suit.suitID]).Do(item =>
+                {
+                    if (item != null)
+                    {
+                        item.hasBeenUnlockedByPlayer = !item.alreadyUnlocked;
+                    }
+                });
+
+                // Sync purchased ship objects.
+                placeableObjects.Select(furnitureID => ___unlockablesList.unlockables[furnitureID]).Do(item =>
+                {
+                    if (item != null)
+                    {
+                        item.hasBeenUnlockedByPlayer = !item.alreadyUnlocked;
+                    }
+                });
+
+                // Sync ship objects in storage.
+                storedItems.Select(storedID => ___unlockablesList.unlockables[storedID]).Do(item =>
+                {
+                    if (item != null)
+                    {
+                        item.hasBeenUnlockedByPlayer = !item.alreadyUnlocked;
+                        item.inStorage = true;
+                    }
+                });
+
+                // Sync cozy lights (neither suit nor placeable object).
+                if (Object.FindObjectOfType<CozyLights>() != null)
+                {
+                    UnlockableItem? item = ___unlockablesList.unlockables.Find(item => string.CompareOrdinal(item.unlockableName, "Cozy lights") == 0);
+                    if (item != null)
+                    {
+                        item.hasBeenUnlockedByPlayer = true;
+                    }
+                }
+
+                // Unlockables should be properly synced at this point.
+                UnlockablesSynced = true;
+            }
+
+            // Manually trigger a store rotation on the local client.
             Plugin.Terminal?.RotateShipDecorSelection();
         }