+++ StoreRotationConfig/Patches/UnlockShipObjectPatches.cs
@@ -1,98 +1,77 @@
 using HarmonyLib;
-using Unity.Netcode;
 
 using static StoreRotationConfig.Api.RotationItemsAPI;
 
 namespace StoreRotationConfig.Patches
 {
     /// <summary>
-    ///     Patches for removing purchased items from both current and future store rotations.
+    ///     Patches for removing purchased items from current and future store rotations.
     /// </summary>
     [HarmonyPatch(typeof(StartOfRound))]
     internal class UnlockShipObjectPatches
     {
-        [HarmonyPatch("UnlockShipObject", typeof(int))]
-        [HarmonyPrefix]
-        private static void UnlockShipObjectPre(int unlockableID)
-        {
-            // Return if the local game instance is not hosting the server, or if 'stockPurchased' setting is enabled.
-            if ((!NetworkManager.Singleton.IsHost && !NetworkManager.Singleton.IsServer) || (Plugin.Settings?.STOCK_PURCHASED.Value).GetValueOrDefault(true))
-            {
-                return;
-            }
-
-            // Attempt to remove item from the store rotation.
-            Plugin.StaticLogger?.LogDebug($"Attempting to remove unlockable #{unlockableID} on server...");
-            RemoveItemFromRotation(unlockableID);
-        }
-
-        [HarmonyPatch(nameof(StartOfRound.BuyShipUnlockableClientRpc), typeof(int), typeof(int))]
-        [HarmonyPrefix]
-        private static void BuyShipUnlockableClientPre(int newGroupCreditsAmount, int unlockableID = -1)
-        {
-            // Return if local game instance is hosting the server, or if the 'stockPurchased' setting is enabled.
-            if (NetworkManager.Singleton.IsHost || NetworkManager.Singleton.IsServer || (Plugin.Settings?.STOCK_PURCHASED.Value).GetValueOrDefault(true))
-            {
-                return;
-            }
-
-            // Attempt to remove item from the store rotation.
-            Plugin.StaticLogger?.LogDebug($"Attempting to remove unlockable #{unlockableID} on client...");
-            RemoveItemFromRotation(unlockableID);
-        }
-
         /// <summary>
-        ///     Helper function for removing items from both 'RotateShipDecorSelectionPatch.AllItems' and 'Terminal.ShipDecorSelection' lists.
+        ///     Remove an item from current and future store rotations, and also from the list of permanent items.
         /// </summary>
-        /// <param name="unlockableID">ID of the unlockable item to remove.</param>
-        private static void RemoveItemFromRotation(int unlockableID)
+        /// <param name="item">'UnlockableItem' instance of the item to remove.</param>
+        /// <param name="unlockableID">The ID of the item to remove (only used for logging).</param>
+        private static void RemoveFromRotation(UnlockableItem? item, int unlockableID = -1)
         {
-            // Return if no item was successfully purchased.
-            if (unlockableID == -1)
-            {
-                return;
-            }
-
-            // Obtain item from the list of purchasable items.
-            UnlockableItem item = StartOfRound.Instance.unlockablesList.unlockables[unlockableID];
-
-            // Return if item is not present in the 'UnlockablesList.unlockables' list, OR its shop node does not exist.
+            // Return if item OR its shop node does not exist.
             if (item == null || item.shopSelectionNode == null)
             {
-                Plugin.StaticLogger?.LogWarning($"Unlockable #{unlockableID} and/or its terminal node could not be found.");
+                Plugin.StaticLogger?.LogWarning($"Item #{unlockableID} and/or its terminal node could not be found.");
 
                 return;
             }
 
-            // Return if item has already been purchased (likely a redundant check, but done just in case).
-            if (item.hasBeenUnlockedByPlayer || item.alreadyUnlocked)
-            {
-                Plugin.StaticLogger?.LogWarning($"Item '{item.shopSelectionNode.creatureName}' has already been purchased.");
-
-                return;
-            }
-
-            Plugin.StaticLogger?.LogDebug($"Removing item '{item.shopSelectionNode.creatureName}' from the store rotation...");
-
-            // Attempt to remove item from the 'RotateShipDecorSelectionPatch.AllItems' list.
-            if (UnregisterItem(item))
-            {
-                // Attempt to remove item from the 'Terminal.ShipDecorSelection' list.
-                if (!Plugin.Terminal.ShipDecorSelection.Remove(item.shopSelectionNode))
-                {
-                    Plugin.StaticLogger?.LogWarning($"Item '{item.shopSelectionNode.creatureName}' could not be removed from the store rotation.");
-                }
-            }
-            else
-            {
-                Plugin.StaticLogger?.LogWarning($"Item '{item.shopSelectionNode.creatureName}' could not be found in the list of purchasable items.");
-            }
+            Plugin.StaticLogger?.LogDebug($"Attempting to remove item '{item.unlockableName}' from the store rotation on local client...");
 
             // Attempt to remove item from the 'RotateShipDecorSelectionPatch.PermanentItems' list.
             if (RemovePermanentItem(item))
             {
-                Plugin.StaticLogger?.LogDebug($"Item '{item.shopSelectionNode.creatureName}' was removed from the list of permanent items.");
+                Plugin.StaticLogger?.LogDebug($"Removed item '{item.shopSelectionNode.creatureName}' from the list of permanent items.");
             }
+
+            // Attempt to remove item from the current store rotation.
+            if (Plugin.Terminal?.ShipDecorSelection.Remove(item.shopSelectionNode) == true)
+            {
+                Plugin.StaticLogger?.LogDebug($"Removed item '{item.shopSelectionNode.creatureName}' from the current store rotation.");
+            }
+
+            // Attempt to remove item from future store rotations.
+            if (UnregisterItem(item))
+            {
+                Plugin.StaticLogger?.LogDebug($"Removed item '{item.shopSelectionNode.creatureName}' from future store rotations.");
+            }
+        }
+
+        [HarmonyPatch("UnlockShipObject")]
+        [HarmonyPrefix]
+        private static void UnlockShipObjectPre(StartOfRound __instance, int unlockableID)
+        {
+            // Return if not running from server, no item was successfully purchased, OR the 'removePurchased' setting is not enabled.
+            if (!__instance.IsHost || unlockableID == -1 || Plugin.Settings?.REMOVE_PURCHASED.Value == false)
+            {
+                return;
+            }
+
+            // Attempt to remove purchased item from store rotation on server.
+            RemoveFromRotation(__instance.unlockablesList?.unlockables[unlockableID], unlockableID);
+        }
+
+        [HarmonyPatch(nameof(StartOfRound.BuyShipUnlockableClientRpc))]
+        [HarmonyPrefix]
+        private static void BuyShipUnlockableClientRpcPre(StartOfRound __instance, int unlockableID = -1)
+        {
+            // Return if running from server, no item was successfully purchased, OR the 'removePurchased' setting is not enabled.
+            if (__instance.IsHost || unlockableID == -1 || Plugin.Settings?.REMOVE_PURCHASED.Value == false)
+            {
+                return;
+            }
+
+            // Attempt to remove purchased item from store rotation on clients.
+            RemoveFromRotation(__instance.unlockablesList?.unlockables[unlockableID], unlockableID);
         }
     }
 }